Update myscript.sh
here the msbar tool is used to incorporate mismatches in the DNA sequence provided(sample 09). So, the limitations of the blast in detecting mismatches can also be detected by providing this command line. the value of is changed in 50 increments the command was run to assess the change of bit score , evalue and the number of mismatches.