Update codes for question 1 to 5
the data included in "gene_expression.tsv" file is accessed and the new column was added to the data set as mean. ten highest mean values were extracted from the data set. finally the data were plotted in pairs plot. here the coding is mainly done for access a particular data base , extract specific data from it and to analyse them making a pairs plot. making pairs plots is really important in analysing the real distribution of data with compared to line or bar chart.