Update myscript.sh
In practical eight, msbar tool is used to incoorporate mismatches in in the provided DNA sequence(sample 09). so, the important data such as, length of the sequence, e-value, number of mismatches and bit score can be anlised within a very short time. using this tool is really important in extract a specific genetic information from the huge data set and to do the analyse. the ability to save the codes in GitHub also helps to use the knowledge for future studies.